+++ ostap/parallel/tests/test_parallel_ipyparallel.py
@@ -9,7 +9,7 @@
 from   itertools                import count    
 from   ostap.utils.progress_bar import progress_bar 
 from   ostap.plotting.canvas    import use_canvas
-from   ostap.utils.utils        import wait 
+from   ostap.utils.timing       import timing 
 import ostap.histos.histos
 import ROOT, time, sys, warnings  
 # =============================================================================
@@ -65,8 +65,8 @@ class MakeHisto(object) :
 
 mh  = MakeHisto  ()
 
-## start 50 jobs, and for each job create the histogram with 1000 entries 
-inputs = 50 * [ 1000 ]
+## start 50 jobs, and for each job create the histogram with 100 entries 
+inputs = 50 * [ 100 ]
 
 # =============================================================================
 ## test parallel processing with ipyparallel
@@ -74,7 +74,7 @@ def test_ipyparallel_function () :
     """Test parallel processnig with ipyparallel
     """
 
-    logger = getLogger ( "ostap.test_ipyparallel_function")    
+    logger = getLogger ( "test_ipyparallel_function")    
     logger.info ('Test job submission with ipyparallel')
     
     if not (3,6)<= sys.version_info :
@@ -89,17 +89,17 @@ def test_ipyparallel_function () :
         return
 
     result = None 
-    with ipp.Cluster() as cluster :
-
-        view    = cluster.load_balanced_view()
+    with ipp.Cluster () as cluster :
         
+        view    = cluster.load_balanced_view()        
         results = view.map_async ( make_histos , zip  ( count () , inputs ) )
         
         for r in progress_bar ( results ) :
             if not result  : result = r
             else           : result.Add ( r )
             
-    with wait ( 3 ) , use_canvas ( 'test_ipyparallel_function' ) : 
+               
+    with use_canvas ( 'test_ipyparallel_function' , wait = 2 ) : 
         logger.info ( "Histogram is %s" % result.dump ( 80 , 20 ) )
         logger.info ( "Entries  %s/%s" % ( result.GetEntries() , sum ( inputs ) ) ) 
         result.draw (   ) 
@@ -112,7 +112,7 @@ def test_ipyparallel_callable () :
     """Test parallel processnig with ipyparallel
     """
 
-    logger = getLogger ( "ostap.test_ipyparallel_callable")    
+    logger = getLogger ( "test_ipyparallel_callable")    
     logger.info ('Test job submission with ipyparallel')
     
     if not (3,6)<= sys.version_info :
@@ -129,10 +129,10 @@ def test_ipyparallel_callable () :
     if not dill :
         logger.error ( "dill is not available" ) 
         return
-        
-    result = None 
-    with ipp.Cluster() as cluster :
 
+    result = None 
+    with ipp.Cluster () as cluster :
+        
         ##view    = cluster.load_balanced_view()
         view    = cluster[:] 
         view.use_dill()
@@ -142,8 +142,8 @@ def test_ipyparallel_callable () :
         for r in progress_bar ( results ) :
             if not result  : result = r
             else           : result.Add ( r )
-            
-    with wait ( 3 ) , use_canvas ( 'test_ipyparallel_function' ) : 
+                            
+    with use_canvas ( 'test_ipyparallel_function' , wait = 2 ) : 
         logger.info ( "Histogram is %s" % result.dump ( 80 , 20 ) )
         logger.info ( "Entries  %s/%s" % ( result.GetEntries() , sum ( inputs ) ) ) 
         result.draw (   ) 
@@ -153,9 +153,11 @@ def test_ipyparallel_callable () :
     
 # =============================================================================
 if '__main__' == __name__ :
-
-    test_ipyparallel_function () 
-    test_ipyparallel_callable () 
+    
+    with timing ('test_ipyrallel_function' , logger =logger ) :  
+        test_ipyparallel_function () 
+    with timing ('test_ipyrallel_callable' , logger =logger ) :  
+        test_ipyparallel_callable () 
         
 # =============================================================================
 ##                                                                      The END