+++ ostap/parallel/tests/test_parallel_parallel_gaudi.py
@@ -58,8 +58,8 @@ def merge_histos  ( h1 , h2 ) :
     return h1
 
 # =============================================================================
-## start 5 jobs, and for each job create the histogram with 100 entries 
-inputs = 5 * [ 100 ] 
+## start 50 jobs, and for each job create the histogram with 100 entries 
+inputs = 50 * [ 100 ] 
 
 # ==============================================================================
 ## simple task to create and fill historgam 
@@ -85,7 +85,7 @@ class HTask(Task) :
 def test_parallel_gaudi_mp_bare ( ) :
     """Test parallel processnig with parallel_gaudi (bare interface) 
     """
-    logger  = getLogger ("ostap.test_parallel_gaudi_mp_bare")
+    logger  = getLogger ("test_parallel_gaudi_mp_bare")
     if not WorkManager :
         logger.error ("Failure to import WorkManager")
         return
@@ -123,7 +123,7 @@ def test_parallel_gaudi_mp_bare ( ) :
 def test_parallel_gaudi_mp_task ( ) :
     """Test parallel processnig with parallel_gaudi (task interface) 
     """
-    logger  = getLogger ("ostap.test_parallel_gaudi_mp_task")
+    logger  = getLogger ("test_parallel_gaudi_mp_task")
     if not WorkManager :
         logger.error ("Failure to import WorkManager")
         return
@@ -159,7 +159,7 @@ def test_parallel_gaudi_mp_task ( ) :
 def test_parallel_gaudi_mp_func ( ) :
     """Test parallel processnig with parallel_gaudi (func interface) 
     """
-    logger  = getLogger ("ostap.test_parallel_gaudi_mp_task")
+    logger  = getLogger ("test_parallel_gaudi_mp_task")
     if not WorkManager :
         logger.error ("Failure to import WorkManager")
         return
@@ -192,7 +192,7 @@ def test_parallel_gaudi_mp_func ( ) :
 def test_parallel_gaudi_mp_generic ( ) :
     """Test parallel processnig with parallel_gaudi (use generic task)
     """
-    logger  = getLogger ("ostap.test_parallel_gaudi_mp_generic")
+    logger  = getLogger ("test_parallel_gaudi_mp_generic")
     if not WorkManager :
         logger.error ("Failure to import WorkManager")
         return